% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTable.R
\name{stockTable_query}
\alias{stockTable_query}
\title{嘉好即时库存更新报表}
\usage{
stockTable_query(
  token,
  FMaterialId,
  companyName,
  Flot,
  FBillNO,
  Fsalesman,
  FCustName
)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
嘉好即时库存更新报表
}
\examples{
stockTable_query('C0426D23-1927-4314-8736-A74B2EF7A039','01.3889U.904','江苏嘉好热熔胶股份有限公司','R-230809013-A03B-5202U-AD(1-)(印尼)','SCDD-100-20230807-0002','谢纪伟','PT.ANEKA MITRA GEMILANG')
}
